控制器方法重构
控制器方法重构
