+++ src/org/ramer/diary/handler/user/ModifyPassword.java
@@ -1,0 +1,91 @@
+package org.ramer.diary.handler.user;
+
+import java.io.IOException;
+import java.util.Map;
+
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import org.ramer.diary.constant.PageConstant;
+import org.ramer.diary.domain.Topic;
+import org.ramer.diary.domain.User;
+import org.ramer.diary.exception.SystemWrongException;
+import org.ramer.diary.exception.UserNotLoginException;
+import org.ramer.diary.service.UserService;
+import org.ramer.diary.util.Encrypt;
+import org.ramer.diary.util.UserUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttributes;
+
+/**
+ *修改密码.
+ * @author ramer
+ *
+ */
+@SessionAttributes(value = { "user", "topics", }, types = { User.class, Topic.class })
+@Controller
+public class ModifyPassword {
+
+  @Autowired
+  UserService userService;
+  //全局成功页面
+  final String SUCCESS = PageConstant.SUCCESS.toString();
+
+  /**
+   * 重定向到修改密码页面.
+   *
+   * @param map the map
+   * @param session the session
+   * @return 引导到修改密码页面
+   */
+  @RequestMapping(value = "/user/forwardModifyPassword", method = RequestMethod.GET)
+  public String forwardModifyPassword(Map<String, Object> map, HttpSession session) {
+    if (!UserUtils.checkLogin(session)) {
+      throw new UserNotLoginException("您还未登录或登录已过期");
+    }
+    System.out.println("引导到修改用户密码页面");
+    return "modifyPass";
+  }
+
+  /**
+   * 修改密码.
+   *
+   * @param oldPassword 原始密码
+   * @param newPassword 新密码
+   * @param user 用户
+   * @param map the map
+   * @param response the response
+   * @param session the session
+   * @return 密码修改成功: 返回个人主页,失败: 返回密码修改页面
+   * @throws IOException Signals that an I/O exception has occurred.
+   */
+  @RequestMapping("/user/modifyPassword")
+  public String modifyPassword(@RequestParam("oldPassword") String oldPassword,
+      @RequestParam("newPassword") String newPassword, User user, Map<String, Object> map,
+      HttpServletResponse response, HttpSession session) throws IOException {
+    if (!Encrypt.execEncrypt(oldPassword).equals(user.getPassword())) {
+      System.out.println("原始密码错误");
+      session.setAttribute("error_modifyPass", "原始密码错误");
+      return "redirect:/user/forwardModifyPassword";
+    }
+    if (oldPassword.equals(newPassword)) {
+      System.out.println("密码未改变");
+      session.setAttribute("error_modifyPass", "");
+      session.setAttribute("succMessage", "密码修改成功");
+      return SUCCESS;
+    }
+    user.setPassword(Encrypt.execEncrypt(newPassword));
+    user = userService.newOrUpdate(user);
+    if (user == null) {
+      throw new SystemWrongException();
+    }
+    session.setAttribute("error_modifyPass", "");
+    session.setAttribute("succMessage", "密码修改成功");
+    return SUCCESS;
+  }
+
+}
